I did it!
I know this ain't proffessional so just read this.

I had to add like a million lines to Firebase_Service so don't touch it unless you dare.

I also had to change some things in the borrowpage and lending page but not alot of stuff.

Basically though it has to have certain data values to appear which is as following:

It needs a created date, item name, and a user id.
the lending page takes a ownerref or owner id. this is the id of who is lending the item out

the borrow page takes the borrowerref or borrower id. this is the id of who is borrowing.

this means to truly test this feature we need two users to lend and borrow from each other.

I will hard code the create to make lending easy borrowing program will happen in the main page as thats probably the easiest for the demo this thursday.

Thats all from me so goodnight!