Got add Items to work and I give up
Add items, adds a new item.

I give up on trying to get borrow items to work properly. right now it calls it when its not equal to.

the logic makes total sense, if BorrowRef is = to the current user it should appear. it works just fine for the lender page but it just wont work for borrow idk.